backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
geoalchemy2
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
shapely
python-dotenv
alembic
cors
rasterio
numpy
scipy
fiona
pyogrio
reportlab
xlsxwriter
